requirements.txt
Pillow >= 5.3
joblib >= 0.13
numpy >= 1.15

Flask >= 1.1
